fix: prevent potential IRM underflow

+++ src/InterestRateModels/IRMLinearKink.sol
@@ -19,7 +19,7 @@ contract IRMLinearKink is IIRM {
     /// @notice Utilization at which the slope of the interest rate function changes. In type(uint32).max scale.
     uint256 public immutable kink;
 
-    constructor(uint256 baseRate_, uint256 slope1_, uint256 slope2_, uint256 kink_) {
+    constructor(uint256 baseRate_, uint256 slope1_, uint256 slope2_, uint32 kink_) {
         baseRate = baseRate_;
         slope1 = slope1_;
         slope2 = slope2_;
